refactor: simplify src/daemon/rpc.rs
Split the monolithic 3361-line rpc.rs into focused domain modules:

- rpc.rs (815 lines): Thin dispatch routing table
- rpc_channel.rs (402 lines): channel.post, channel.read
- rpc_coworker.rs (437 lines): coworker lifecycle and state
- rpc_tasks.rs (965 lines): task CRUD, clusterer integration
- rpc_sessions.rs (417 lines): attach/detach for debugging
- rpc_handlers.rs (453 lines): status, reminders, insight, headless

Benefits:
- Clear separation of concerns by domain
- Easier to locate handler logic
- Tests stay with their handlers
- Main dispatch table shows full API surface at a glance

All tests pass. No behavioral changes.
