+++ src/daemon/mod.rs
@@ -18,6 +18,12 @@ pub mod helpers;
 mod pr;
 mod rpc;
 mod rpc_auth;
+mod rpc_channel;
+mod rpc_coworker;
+mod rpc_handlers;
+mod rpc_kanban;
+mod rpc_sessions;
+mod rpc_tasks;
 pub(crate) mod sessions;
 pub(crate) mod snapshot;
 mod specialized;
@@ -497,7 +503,7 @@ pub(crate) struct DaemonState {
     /// Stores the full kanban GraphQL response (PRs, merged PRs, repos) keyed by
     /// a hash of the repo paths. Integrated into DaemonState (rather than a global
     /// static) so the daemon can inspect and clean it up alongside other caches.
-    kanban_cache: rpc::KanbanCache,
+    kanban_cache: rpc_kanban::KanbanCache,
     /// Draining mode flag - when true, daemon stops assigning new tasks to coworkers.
     ///
     /// Set via `daemon.enter-drain` RPC when `midtown restart` is called without --force.
@@ -695,7 +701,7 @@ impl DaemonState {
             attached_coworkers: std::sync::Mutex::new(HashSet::new()),
             session_manager: sessions::SessionManager::new(session_manager_repo_name),
             rpc_response_cache: Mutex::new(HashMap::new()),
-            kanban_cache: rpc::KanbanCache::new(),
+            kanban_cache: rpc_kanban::KanbanCache::new(),
             draining: std::sync::atomic::AtomicBool::new(false),
             restart_requested: std::sync::atomic::AtomicBool::new(false),
             shutdown_tx,
@@ -2092,7 +2098,7 @@ pub async fn run(config: DaemonConfig) -> crate::Result<DaemonExitStatus> {
             } => {
                 let content = &mobile_post.content;
                 let sender = state.user_display_name.as_deref().unwrap_or("user");
-                rpc::handle_channel_post(
+                rpc_channel::handle_channel_post(
                     RequestId::Null,
                     sender,
                     content,
